day3/requirements.txt
numpy==2.4.4
google-genai
python-dotenv
